Rename: 토론방 관련 controller 하나의 폴더로 관리하도록 변경
